refactor: Modal 디렉토리를 lib/components/Modal로 이동
