+++ components/src/lib/components/Modal/CloseButtonIcon/index.tsx
