+++ components/src/lib/index.tsx
@@ -1,1 +1,1 @@
-export { default as Modal } from './Modal';
+export { default as Modal } from './components/Modal';
